simplify redirect logic
